// shared-schemas/src/types.d.ts
export interface Point2D {
    x_mm: number;
    y_mm: number;
}
export interface Point3D {
    x_mm: number;
    y_mm: number;
    z_mm: number;
}
export interface Vector2D {
    x: number;
    y: number;
}
export interface Rotation3D {
    x_deg: number;
    y_deg: number;
    z_deg: number;
}
export interface Vertex extends Point2D {
    id: string;
    index: number;
}